requirements.txt
torch==2.2.1
torchvision==0.17.1
numpy==1.26.4
pillow==10.2.0
uvicorn
fastapi==0.110.0
aiofiles
python-multipart
scikit-image==0.22.0
gdown
opencv-python==4.9.0.80
scipy==1.12.0
pandas==2.2.2
requests==2.31.0
